+++ solidity/contracts/converter/interfaces/ITypedConverterAnchorFactory.sol
@@ -5,6 +5,6 @@ import "./IConverterAnchor.sol";
     Typed Converter Anchor Factory interface
 */
 contract ITypedConverterAnchorFactory {
-    function converterType() public pure returns (uint8);
+    function converterType() public pure returns (uint16);
     function createAnchor(string _name, string _symbol, uint8 _decimals) public returns (IConverterAnchor);
 }
