+++ solidity/contracts/converter/interfaces/ITypedConverterFactory.sol
@@ -7,6 +7,6 @@ import "../../utility/interfaces/IContractRegistry.sol";
     Typed Converter Factory interface
 */
 contract ITypedConverterFactory {
-    function converterType() public pure returns (uint8);
+    function converterType() public pure returns (uint16);
     function createConverter(IConverterAnchor _anchor, IContractRegistry _registry, uint32 _maxConversionFee) public returns (IConverter);
 }
